refactor(api): split repository god-files into per-module files
Split drizzle.ts (611 lines, 7 classes) and in-memory.ts (442 lines, 7 classes)
into dedicated files under drizzle/ and in-memory/ directories. Shared column
definitions and helpers extracted to drizzle/shared.ts. Barrel index.ts files
preserve existing import paths so the composition root is unchanged.

No logic changes -- pure mechanical extraction.
